Implement CBOR and MessagePack serialization support in the Rust backend
